[fei5585.1.hooksrefresh] Refactoring hooks and general API
