+++ __docs__/wonder-blocks-timing/use-timeout.stories.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 
 import {View} from "@khanacademy/wonder-blocks-core";
-import Button from "@khanacademy/wonder-blocks-button";
 
 import {useTimeout} from "@khanacademy/wonder-blocks-timing";
 
@@ -17,18 +16,13 @@ export default {
 
 export const BasicUsage = () => {
     const [callCount, setCallCount] = React.useState(0);
-    const [active, setActive] = React.useState(false);
     const callback = React.useCallback(() => {
         setCallCount((callCount) => callCount + 1);
     }, []);
-    useTimeout(callback, 1000, active);
+    useTimeout(callback, 1000);
     return (
         <View>
             <View>callCount = {callCount}</View>
-            <View>active = {active.toString()}</View>
-            <Button onClick={() => setActive(!active)} style={{width: 200}}>
-                Toggle active
-            </Button>
         </View>
     );
 };
