+++ packages/wonder-blocks-timing/src/util/interval.ts
@@ -1,9 +1,6 @@
-import {
-    SchedulePolicy as SchedulePolicies,
-    ClearPolicy as ClearPolicies,
-} from "./policies";
+import {SchedulePolicy, ClearPolicy} from "./policies";
 
-import type {IInterval, SchedulePolicy, ClearPolicy} from "./types";
+import type {IInterval} from "./types";
 
 /**
  * Encapsulates everything associated with calling setInterval/clearInterval,
@@ -23,10 +20,10 @@ export default class Interval implements IInterval {
      * Creates an interval that will invoke the given action after
      * the given period. The interval does not start until set is called.
      *
-     * @param {() => mixed} action The action to be invoked each time the
+     * @param action The action to be invoked each time the
      * interval period has passed.
-     * @param {number} intervalMs The interval period.
-     * @param {SchedulePolicy} [schedulePolicy] When SchedulePolicy.Immediately,
+     * @param intervalMs The interval period.
+     * @param [schedulePolicy] When SchedulePolicy.Immediately,
      * the interval is set immediately on instantiation; otherwise, `set` must be
      * called to set the interval.
      * Defaults to `SchedulePolicy.Immediately`.
@@ -35,7 +32,7 @@ export default class Interval implements IInterval {
     constructor(
         action: () => unknown,
         intervalMs: number,
-        schedulePolicy: SchedulePolicy = SchedulePolicies.Immediately,
+        schedulePolicy: SchedulePolicy = SchedulePolicy.Immediately,
     ) {
         if (typeof action !== "function") {
             throw new Error("Action must be a function");
@@ -48,7 +45,7 @@ export default class Interval implements IInterval {
         this._action = action;
         this._intervalMs = intervalMs;
 
-        if (schedulePolicy === SchedulePolicies.Immediately) {
+        if (schedulePolicy === SchedulePolicy.Immediately) {
             this.set();
         }
     }
@@ -56,7 +53,7 @@ export default class Interval implements IInterval {
     /**
      * Determine if the interval is active or not.
      *
-     * @returns {boolean} true if the interval is active, otherwise false.
+     * @returns true if the interval is active, otherwise false.
      * @memberof Interval
      */
     get isSet(): boolean {
@@ -73,7 +70,7 @@ export default class Interval implements IInterval {
      */
     set(): void {
         if (this.isSet) {
-            this.clear(ClearPolicies.Cancel);
+            this.clear(ClearPolicy.Cancel);
         }
         this._intervalId = setInterval(() => this._action(), this._intervalMs);
     }
@@ -84,22 +81,21 @@ export default class Interval implements IInterval {
      * If the interval is active, this cancels that interval. If no interval is
      * pending, this does nothing.
      *
-     * @param {ClearPolicy} [policy] When ClearPolicy.Resolve, if the request
+     * @param [policy] When ClearPolicy.Resolve, if the request
      * was set when called, the request action is invoked after cancelling
      * the request; otherwise, the pending action is cancelled.
      * Defaults to `ClearPolicy.Cancel`.
      *
-     * @returns {void}
      * @memberof Interval
      */
-    clear(policy: ClearPolicy = ClearPolicies.Cancel): void {
+    clear(policy: ClearPolicy = ClearPolicy.Cancel): void {
         const intervalId = this._intervalId;
         this._intervalId = null;
         if (intervalId == null) {
             return;
         }
         clearInterval(intervalId);
-        if (policy === ClearPolicies.Resolve) {
+        if (policy === ClearPolicy.Resolve) {
             this._action();
         }
     }
