+++ packages/wonder-blocks-timing/tsconfig-build.json
@@ -3,9 +3,7 @@
     "extends": "../tsconfig-shared.json",
     "compilerOptions": {
         "outDir": "./dist",
-        "rootDir": "src",
+        "rootDir": "src"
     },
-    "references": [
-        {"path": "../wonder-blocks-button/tsconfig-build.json"},
-    ]
+    "references": []
 }
